src/requirements.txt
# fastapi
# uvicorn[standard]
# pydantic
# python-multipart
# celery
# redis
# flower
# pydantic
# python-dotenv
# celery
# torch
# numpy

torch
torchvision
numpy
pandas
geopandas
scikit-learn
colour-science
shapely
openslide-python
opencv-python-headless
opencv-contrib-python-headless
Pillow
matplotlib
seaborn
tifffile
multiprocess
patchify
pyjwt
scikit-image
wandb
fastapi
uvicorn[standard]
pydantic
python-multipart
celery
redis
flower
python-dotenv
tqdm
sqlalchemy
psycopg2-binary

